+++ src/main/java/com/advogado/freelancer/useCases/usuarios/UsuarioBusiness.java
@@ -1,6 +1,7 @@
 package com.advogado.freelancer.useCases.usuarios;
 import com.advogado.freelancer.entities.Usuario;
 import com.advogado.freelancer.frameWork.utils.SenacException;
+import com.advogado.freelancer.useCases.usuarios.domanis.UsuarioLoginDTO;
 import com.advogado.freelancer.useCases.usuarios.domanis.UsuarioRequestDom;
 import com.advogado.freelancer.useCases.usuarios.domanis.UsuarioResponseDom;
 import java.util.List;
