+++ src/main/java/com/advogado/freelancer/useCases/usuarios/impl/UsuarioServiceImpl.java
@@ -2,6 +2,7 @@ package com.advogado.freelancer.useCases.usuarios.impl;
 import com.advogado.freelancer.entities.Usuario;
 import com.advogado.freelancer.frameWork.utils.SenacException;
 import com.advogado.freelancer.useCases.usuarios.UsuarioService;
+import com.advogado.freelancer.useCases.usuarios.domanis.UsuarioLoginDTO;
 import com.advogado.freelancer.useCases.usuarios.domanis.UsuarioRequestDom;
 import com.advogado.freelancer.useCases.usuarios.domanis.UsuarioResponseDom;
 import org.springframework.beans.factory.annotation.Autowired;
